Tabulator i backend

+++ geotech-pro-backend/geotech_backend/geotech/serializers.py
@@ -6,7 +6,7 @@ from django.contrib.auth.models import User
 class CptDataSerializer(serializers.ModelSerializer):
     class Meta:
         model = CptData
-        fields = ['id', 'depth', 'qc', 'fs', 'u']
+        fields = ['id', 'depth', 'qc', 'fs', 'u2']
 
 class CptTestSerializer(serializers.ModelSerializer):
     data = CptDataSerializer(many=True)
@@ -37,41 +37,58 @@ class LayerSerializer(serializers.ModelSerializer):
         fields = ['id', 'name', 'depth', 'unit_weight', 'cohesion', 'friction_angle', 'compressibility', 'permeability', 'cpt_data']
 
 class GeotechnicalModelSerializer(serializers.ModelSerializer):
-    layers = LayerSerializer(many=True)
-    cpt_tests = CptTestSerializer(many=True)
+    layers = LayerSerializer(many=True, required=False)
+    cpt_tests = CptTestSerializer(many=True, required=False)
     user = serializers.PrimaryKeyRelatedField(read_only=True)
+    project_id = serializers.IntegerField(source='project.id', read_only=True)
 
     class Meta:
         model = GeotechnicalModel
-        fields = ['id', 'user', 'name', 'npv', 'npv_max', 'layers', 'cpt_tests']
+        fields = ['id', 'user', 'name', 'npv', 'npv_max', 'layers', 'cpt_tests', 'project', 'project_id']
+
+    def to_representation(self, instance):
+        data = super().to_representation(instance)
+        if not instance.layers.exists():
+            data['layers'] = []
+        if not instance.cpt_tests.exists():
+            data['cpt_tests'] = []
+        return data
 
     def create(self, validated_data):
-        layers_data = validated_data.pop('layers')
-        cpt_tests_data = validated_data.pop('cpt_tests')
+        layers_data = validated_data.pop('layers', [])
+        cpt_tests_data = validated_data.pop('cpt_tests', [])
         model = GeotechnicalModel.objects.create(user=self.context['request'].user, **validated_data)
+        
         for layer_data in layers_data:
             Layer.objects.create(model=model, **layer_data)
+            
         for cpt_test_data in cpt_tests_data:
-            data = cpt_test_data.pop('data')
+            data = cpt_test_data.pop('data', [])
             cpt_test = CptTest.objects.create(model=model, **cpt_test_data)
             for item in data:
                 CptData.objects.create(cpt_test=cpt_test, **item)
         return model
 
     def update(self, instance, validated_data):
-        layers_data = validated_data.pop('layers')
-        cpt_tests_data = validated_data.pop('cpt_tests')
+        # Ažuriramo osnovne podatke modela
         instance.name = validated_data.get('name', instance.name)
         instance.npv = validated_data.get('npv', instance.npv)
         instance.npv_max = validated_data.get('npv_max', instance.npv_max)
         instance.save()
-        instance.layers.all().delete()
-        instance.cpt_tests.all().delete()
-        for layer_data in layers_data:
-            Layer.objects.create(model=instance, **layer_data)
-        for cpt_test_data in cpt_tests_data:
-            data = cpt_test_data.pop('data')
-            cpt_test = CptTest.objects.create(model=instance, **cpt_test_data)
-            for item in data:
-                CptData.objects.create(cpt_test=cpt_test, **item)
+
+        # Ažuriramo slojeve samo ako su prisutni u podacima
+        if 'layers' in validated_data:
+            instance.layers.all().delete()
+            for layer_data in validated_data['layers']:
+                Layer.objects.create(model=instance, **layer_data)
+
+        # Ažuriramo CPT testove samo ako su prisutni u podacima
+        if 'cpt_tests' in validated_data:
+            instance.cpt_tests.all().delete()
+            for cpt_test_data in validated_data['cpt_tests']:
+                data = cpt_test_data.pop('data', [])
+                cpt_test = CptTest.objects.create(model=instance, **cpt_test_data)
+                for item in data:
+                    CptData.objects.create(cpt_test=cpt_test, **item)
+
         return instance
